refactor: discovery service -> discovery server 로 이름 수정 - 기타 자잘한 수정
