+++ discovery-server/build.gradle
@@ -13,6 +13,10 @@ java {
     }
 }
 
+ext {
+    set('springCloudVersion', "2025.0.0")
+}
+
 repositories {
     mavenCentral()
 }
@@ -24,6 +28,11 @@ dependencies {
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 }
 
+dependencyManagement {
+    imports {
+        mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
+    }
+}
 
 tasks.named('test') {
     useJUnitPlatform()
