+++ discovery-server/src/main/java/com/shipflow/discoveryserver/DiscoveryserverApplication.java
@@ -1,4 +1,4 @@
-package com.shipflow.discoveryservice;
+package com.shipflow.discoveryserver;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,8 +6,8 @@ import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
 
 @SpringBootApplication
 @EnableEurekaServer
-public class DiscoveryserviceApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(DiscoveryserviceApplication.class, args);
-    }
+public class DiscoveryserverApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(DiscoveryserverApplication.class, args);
+	}
 }
