+++ discovery-server/src/main/resources/application.yaml
@@ -1,6 +1,6 @@
 spring:
   application:
-    name: discoveryservice
+    name: discoveryserver
 
 server:
   port: 8761
