+++ discovery-server/src/test/java/com/shipflow/discoveryserver/DiscoveryserviceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.shipflow.discoveryserver;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class discoveryserverApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
