--- discovery-server/src/test/java/com/shipflow/discoveryservice/DiscoveryserviceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.shipflow.discoveryservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DiscoveryserviceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
