+++ template/application.yaml.example
@@ -8,7 +8,7 @@ server:
 eureka:
   client:
     service-url:
-      defaultZone: http://discoveryservice:8761/eureka/
+      defaultZone: http://discoveryserver:8761/eureka/
     registry-fetch-interval-seconds: 5
   instance:
     prefer-ip-address: true
